refactor: architecture cleanup — split CLI, extract quality-retry, add ceremony tests
- Split index.ts (807 lines) into src/cli/commands.ts + src/cli/helpers.ts
- index.ts is now a 40-line thin entry point
- Extract handleQualityFailure to src/ceremonies/quality-retry.ts
- Add tests for refinement, review, retro ceremonies (17 new tests)
